+++ R/Alakazam.R
@@ -110,8 +110,7 @@
 #' @import      methods
 #' @import      utils
 #' @importFrom  airr        read_rearrangement write_rearrangement
-#' @importFrom	ape 		read.tree di2multi reorder.phylo root ladderize
-#' @importFrom  Biostrings  BString extractAt
+#' @importFrom	ape 		read.fastq read.tree di2multi reorder.phylo root ladderize
 #' @importFrom  dplyr       do n desc %>%
 #'                          bind_cols bind_rows combine arrange left_join
 #'                          group_by ungroup
@@ -121,13 +120,11 @@
 #'							right_join rowwise
 #'                          summarize summarize_at
 #'                          transmute rename
-#' @importFrom  GenomicAlignments explodeCigarOps explodeCigarOpLengths                       
 #' @importFrom  igraph      V E graph_from_data_frame as_data_frame as_edgelist 
 #'                          make_graph make_directed_graph make_undirected_graph
 #'                          vertex_attr set_vertex_attr
 #'                          degree shortest_paths all_shortest_paths distances
 #'                          graph_from_adjacency_matrix components groups
-#' @importFrom  IRanges     IRanges
 #' @importFrom  Matrix      sparseMatrix rowSums
 #' @importFrom  progress    progress_bar
 #' @importFrom  readr       read_delim read_tsv write_delim write_tsv cols
@@ -149,6 +146,9 @@
 #' @importFrom  tibble      tibble
 #' @importFrom  tidyr       complete gather
 #' @importFrom  Rcpp        evalCpp
+#' @importFrom  Biostrings         BString extractAt
+#' @importFrom  GenomicAlignments  explodeCigarOps explodeCigarOpLengths                       
+#' @importFrom  IRanges            IRanges
 #' @useDynLib   alakazam, .registration=TRUE
 NULL